Extract editable line from edit component
The edit component contains the functionality to edit a line, which is
useful outside the context of that component. This extracts that
functionality into a shared location.
