+++ src/core/src/components/mod.rs
@@ -2,3 +2,4 @@ pub(crate) mod choice;
 pub(crate) mod confirm;
 pub(crate) mod edit;
 pub(crate) mod help;
+mod shared;
