+++ src/core/src/components/shared/mod.rs
@@ -1,0 +1,3 @@
+mod editable_line;
+
+pub(crate) use editable_line::EditableLine;
